完善管理员登录、权限等功能

+++ src/main/java/com/fortwelve/wechatstore/controller/ManagerController.java
@@ -1,5 +1,6 @@
 package com.fortwelve.wechatstore.controller;
 
+import com.auth0.jwt.interfaces.Claim;
 import com.fortwelve.wechatstore.component.MsgMap;
 import com.fortwelve.wechatstore.controller.ValidatedGroup.addManager;
 import com.fortwelve.wechatstore.controller.ValidatedGroup.updateManager;
@@ -11,32 +12,62 @@ import com.fortwelve.wechatstore.util.JWTUtils;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.data.redis.core.HashOperations;
+import org.springframework.data.redis.core.StringRedisTemplate;
+import org.springframework.data.redis.core.ValueOperations;
+import org.springframework.util.DigestUtils;
 import org.springframework.validation.BindingResult;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.util.HashMap;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
+import java.util.concurrent.TimeUnit;
 
 @Slf4j
-@CrossOrigin
+//@CrossOrigin
 @RestController
 @RequestMapping("/manager")
 public class ManagerController {
-
-    @Autowired
-    ManagerService managerService;
 
     @Value("${JWTUtils.manager.signature}")
     private String managerSignature;
     @Value("${JWTUtils.manager.minute}")
     private int managerMinute;
+    @Value("${Manager.codeTTL}")
+    private long codeTTL;
+
+    @Autowired
+    ManagerService managerService;
+    @Autowired
+    StringRedisTemplate stringRedisTemplate;
+
+    @GetMapping("/verify")
+    public Object verify(@RequestParam String username){
+        MsgMap msg = new MsgMap();
+        try{
+            String randomCode = JWTUtils.getRandomString(4);
+            //放入redis缓存
+            ValueOperations<String,String> valueOperations = stringRedisTemplate.opsForValue();
+            valueOperations.set("verify_code_"+username,randomCode,codeTTL, TimeUnit.SECONDS);
+
+            msg.put("username",username);
+            msg.put("code",randomCode);
+            msg.setMeta("操作成功。",200);
+        }catch (Exception e){
+            e.printStackTrace();
+            log.info(e.getMessage());
+            msg.setMeta("服务器出错。",500);
+        }
+        return msg;
+    }
 
     @PostMapping("/login")
-    public Object login(@RequestParam String username, @RequestParam String password, HttpServletResponse response){
+    public Object login(@RequestParam String username, @RequestParam String data, HttpServletResponse response){
         MsgMap msg = new MsgMap();
         try {
             Manager manager = managerService.getManagerByManagerName(username);
@@ -44,7 +75,18 @@ public class ManagerController {
                 msg.setMeta("管理员不存在。",621);
                 return msg;
             }
-            if(!password.equals(manager.getManager_password())){
+            //从redis缓存中取出code并移除缓存
+            ValueOperations<String,String> valueOperations = stringRedisTemplate.opsForValue();
+            String code = valueOperations.get("verify_code_"+username);
+            valueOperations.getOperations().delete("verify_code_"+username);
+            if(null == code){
+                msg.setMeta("用户未验证或code过期。",624);
+                return msg;
+            }
+            String beforeMD5=manager.getManager_name()+manager.getManager_password()+code;
+            String md5Str = DigestUtils.md5DigestAsHex(beforeMD5.getBytes("utf-8")).toUpperCase();
+
+            if(!md5Str.equals(data.toUpperCase())){
                 msg.setMeta("管理员密码不正确。",622);
                 return msg;
             }
@@ -54,7 +96,7 @@ public class ManagerController {
             Map<String,String> tokenMap = new HashMap<>();
             tokenMap.put("id",String.valueOf(manager.getId()));
             tokenMap.put("username",String.valueOf(manager.getManager_name()));
-            tokenMap.put("roleid",String.valueOf(managerRole.getRole_id()));
+            tokenMap.put("role",managerRole.getRole_name());
             String token = JWTUtils.getToken(tokenMap,managerSignature,managerMinute);
             //设置token
             response.setHeader("token",token);
@@ -77,7 +119,7 @@ public class ManagerController {
         return msg;
     }
     @PostMapping("/addManager")
-    public Object addManager(@Validated(addManager.class) ManagerDTO managerDTO, BindingResult result, HttpServletResponse response){
+    public Object addManager(@Validated(addManager.class) ManagerDTO managerDTO, BindingResult result, HttpServletRequest request){
         MsgMap msg = new MsgMap();
 
         try {
@@ -86,6 +128,19 @@ public class ManagerController {
                 msg.setMeta(result.getFieldError().getDefaultMessage(),701);
                 return msg;
             }
+            String token = request.getHeader("token");
+            Map<String, Claim> tokenMap = JWTUtils.decode(token,managerSignature);
+            String role = tokenMap.get("role").asString();
+            if(!role.equals("超级管理员")){
+                msg.setMeta("没有权限操作。",611);
+                return msg;
+            }
+            Manager isExist=managerService.getManagerByManagerName(managerDTO.getUsername());
+            if(isExist != null){
+                msg.setMeta("管理员已经存在。",623);
+                return msg;
+            }
+
             int id=0;
             boolean find=false;
             List<ManagerRole> managerRoles = managerService.getAllManagerRole();
@@ -132,10 +187,15 @@ public class ManagerController {
         return msg;
     }
     @PostMapping("/updateManager")
-    public Object updateManager(@Validated(updateManager.class) ManagerDTO managerDTO, BindingResult result, HttpServletResponse response){
+    public Object updateManager(@Validated(updateManager.class) ManagerDTO managerDTO, BindingResult result, HttpServletRequest request){
         MsgMap msg = new MsgMap();
 
         try{
+            String token = request.getHeader("token");
+            Map<String,Claim> tokenMap = JWTUtils.decode(token,managerSignature);
+            String idstr = tokenMap.get("id").asString();
+            String userName = tokenMap.get("username").asString();
+            String role = tokenMap.get("role").asString();
             if (result.hasErrors()){
                 msg.setMeta(result.getFieldError().getDefaultMessage(),701);
                 return msg;
@@ -145,6 +205,16 @@ public class ManagerController {
                 msg.setMeta("管理员不存在。",621);
                 return msg;
             }
+            if(!role.equals("超级管理员")){
+                //如果不是超级管理员，就判断修改的管理员信息是否是本人
+                if(!userName.equals(manager.getManager_name())){
+                    //既不是超级管理员，又不是修改自己信息，就不允许
+                    msg.setMeta("没有权限修改他人信息。",611);
+                    return msg;
+                }
+            }
+            //是超级管理员，或者不是超级管理员但修改的是自己信息，才能走到这里
+
             //管理员用户名不可以修改
 //            manager.setManager_name(managerDTO.getUsername());
             if(managerDTO.getPassword() != null ){
@@ -163,22 +233,35 @@ public class ManagerController {
                 manager.setSex(managerDTO.getSex());
             }
             if(managerDTO.getRole() != null ){
-                int id=0;
-                boolean find=false;
+                ManagerRole managerRole=null;
                 List<ManagerRole> managerRoles = managerService.getAllManagerRole();
 
                 for(ManagerRole tmp:managerRoles){
                     if(managerDTO.getRole().equals(tmp.getRole_name())){
-                        id=tmp.getRole_id();
-                        find=true;
+                        managerRole=tmp;
                         break;
                     }
                 }
-                if(find == false){
+
+                if(managerRole == null){
                     msg.setMeta("请求不正确：不存在该角色。",701);
                     return msg;
                 }
-                manager.setRole_id(id);
+                if(managerRole.getRole_id()!=manager.getRole_id()){
+                    //角色不一致，修改角色
+                    //超级管理员可以修改任意人信息，但不能修改自己角色
+                    //其他管理员不可以修改角色
+                    if(role.equals("超级管理员")){
+                        if(userName.equals(manager.getManager_name())){
+                            msg.setMeta("超级管理员不能修改自己角色。",611);
+                            return msg;
+                        }
+                    }else {
+                        msg.setMeta("没有权限修改角色。",611);
+                        return msg;
+                    }
+                    manager.setRole_id(managerRole.getRole_id());
+                }
             }
 
             if(managerService.updateManager(manager)==0){
@@ -201,10 +284,27 @@ public class ManagerController {
         return msg;
     }
     @PostMapping("/deleteManager")
-    public Object deleteManager(int id, HttpServletResponse response){
+    public Object deleteManager(@RequestParam int id, HttpServletRequest request){
         MsgMap msg = new MsgMap();
         try{
-
+            //权限判断
+            String token = request.getHeader("token");
+            Map<String,Claim> tokenMap = JWTUtils.decode(token,managerSignature);
+            String role = tokenMap.get("role").asString();
+            if(!role.equals("超级管理员")){
+                msg.setMeta("没有权限删除管理员。",611);
+                return msg;
+            }
+            Manager manager = managerService.getManagerById(id);
+            if(manager == null){
+                msg.setMeta("管理员不存在。",621);
+                return msg;
+            }
+            ManagerRole managerRole = managerService.getManagerRoleById(manager.getId());
+            if(managerRole.getRole_name().equals("超级管理员")){
+                msg.setMeta("没有权限删除超级管理员。",611);
+                return msg;
+            }
             if(managerService.deleteManagerById(id)==0){
                 msg.setMeta("删除失败。",500);
                 return msg;
@@ -234,6 +334,9 @@ public class ManagerController {
                 managerList=managerService.getManagerPage(head,pageSize);
                 msg.put("currentPage",current);
 
+            }
+            for (Manager tmp : managerList){
+                tmp.setManager_password(null);
             }
             msg.put("total",managerList.size());
             msg.put("list",managerList);
